Fix new api
