+++ extensions/podman/tsconfig.json
@@ -9,6 +9,7 @@
     "outDir": "dist",
     "target": "ES6",
     "moduleResolution": "node",
+    "resolveJsonModule": true,
     "skipLibCheck": true,
     "types": [
         "node",
